pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "vibespatial-raster"
version = "0.1.0"
description = "GPU-first raster processing for vibespatial"
readme = "README.md"
requires-python = ">=3.12"
license = "Apache-2.0"
authors = [
    { name = "vibespatial Contributors" },
]
keywords = ["raster", "gpu", "cuda", "gis", "geospatial", "cupy", "nvrtc", "geotiff"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: GIS",
]
dependencies = [
  "vibespatial>=0.1.0",
  "scipy>=1.14",
]

[project.optional-dependencies]
io = [
  "rasterio>=1.4",
]
gpu-io = [
  "nvidia-nvimgcodec-cu12[all]>=0.6.0",
  "cupy-cuda12x>=13",
]

[project.urls]
Homepage = "https://github.com/picard/vibespatial-raster"
Repository = "https://github.com/picard/vibespatial-raster"
Documentation = "https://picard.github.io/vibespatial-raster/"
Issues = "https://github.com/picard/vibespatial-raster/issues"

[dependency-groups]
dev = [
  "furo>=2025.12.19",
  "myst-parser>=5.0.0",
  "pytest>=8.3",
  "pytest-cov>=6.0",
  "ruff>=0.11",
  "rasterio>=1.4",
  "sphinx>=9.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/vibespatial"]

[tool.pytest.ini_options]
addopts = ["-ra"]
testpaths = ["tests"]
markers = [
  "gpu: requires a CUDA-capable NVIDIA GPU",
]
xfail_strict = true

[tool.ruff]
line-length = 100
target-version = "py312"
